requirements.txt
streamlit
openai
pandas
numpy
faiss-cpu
tiktoken
langchain
langchain-community
python-dotenv
supabase
